Moved rgbd feature extractor to designated file
